requirements.txt
streamlit==1.34.0
google-generativeai==0.8.5
chromadb==0.4.24
sentence-transformers==2.7.0
python-dotenv==1.0.1
pdfkit==1.0.0
ffmpeg-python==0.2.0
faster-whisper==0.10.0
torch==2.1.2
